BF: replace std implementation to fix reduction
It seems that we currently have to know about all reduction operations.
Unfortunately, ``arr.std`` calls into the ``_var`` helper function,
which is a reduction operation, but not an array method, so we can't
intercept it.  Reimplement the ``std`` method using our own wrapped
``arr.var`` to avoid the unwrapped call to ``_var``.